refactor(estimates): extract useEstimateTiers sub-hook (Phase 1d)
Mode/tiers/flatRate state and all tier handlers extracted to dedicated
hook. useEstimateForm now orchestrates 4 sub-hooks and is ~350 lines.
Phase 1 complete.
